Move initial hint into own terminalContrib
