+++ src/vs/workbench/contrib/terminal/terminal.all.ts
@@ -22,6 +22,7 @@ import '../terminalContrib/find/browser/terminal.find.contribution.js';
 import '../terminalContrib/chat/browser/terminal.chat.contribution.js';
 import '../terminalContrib/commandGuide/browser/terminal.commandGuide.contribution.js';
 import '../terminalContrib/history/browser/terminal.history.contribution.js';
+import '../terminalContrib/inlineHint/browser/terminal.initialHint.contribution.js';
 import '../terminalContrib/links/browser/terminal.links.contribution.js';
 import '../terminalContrib/zoom/browser/terminal.zoom.contribution.js';
 import '../terminalContrib/stickyScroll/browser/terminal.stickyScroll.contribution.js';
@@ -32,6 +33,5 @@ import '../terminalContrib/resizeDimensionsOverlay/browser/terminal.resizeDimens
 import '../terminalContrib/sendSequence/browser/terminal.sendSequence.contribution.js';
 import '../terminalContrib/sendSignal/browser/terminal.sendSignal.contribution.js';
 import '../terminalContrib/suggest/browser/terminal.suggest.contribution.js';
-import '../terminalContrib/chat/browser/terminal.initialHint.contribution.js';
 import '../terminalContrib/wslRecommendation/browser/terminal.wslRecommendation.contribution.js';
 import '../terminalContrib/voice/browser/terminal.voice.contribution.js';
