+++ src/vs/workbench/contrib/terminalContrib/inlineHint/browser/media/terminalInitialHint.css
